refactor(sdk): extract AI gateway + MCP to _client/_ai_gateway.py
PR 5/N of the SDK refactor split (see #702 WebSocket, #706 Guardian,
#711 Discovery, #713 RFQ for the mixin template). Moves the three
Mastio-side surfaces — AI gateway egress (ADR-017), MCP aggregator
(ADR-007), and tool-call PDP (ADR-029 Phase D) — out of
cullis_sdk/client.py into a dedicated mixin under cullis_sdk/_client/.

What moves:

- cullis_sdk/_client/_ai_gateway.py — new module exposing
  _AiGatewayMixin with:
    chat_completion, list_models                 (# ── ADR-017 ──)
    list_mcp_tools, call_mcp_tool,
      evaluate_tool_call_policy                  (# ── ADR-007 ──)

- cullis_sdk/client.py:
  * CullisClient now inherits
    (_AiGatewayMixin, _DiscoveryMixin, _GuardianMixin, _RfqMixin,
     _WebSocketMixin)
    in alphabetical order (A < D < G < R < W).
  * Five methods and their two section comment blocks are removed.

- cullis_sdk/__init__.py — unchanged. None of these methods were
  re-exported as top-level symbols.

Preserved behavior:

- chat_completion stays a non-streaming endpoint here. The streaming
  SSE wiring lives proxy-side under llm_chat_router; the SDK method
  parses resp.json() exactly as before. No new SSE event types
  (tool_call_* etc.) added in this PR — that is future scope for the
  SPA v0.1 work.
- evaluate_tool_call_policy keeps every fail-deny branch verbatim:
  transport exception → deny, 401/403 → deny, 5xx/non-200 → deny,
  malformed JSON → deny. 404 → allow with "PDP disabled" reason
  (legacy Mastio bridge). The PDP webhook timeout remains whatever
  the host class's _authed_request already enforces (5s per memory).

No logic change — pure code movement.
